Extract NodeTreeDraw
